ops: Consolidate `ExitError`s into methods
`ExitError` should be opaque, only the methods it exposes can be used
to create values. This was ignored by some of the code in `ops`,
because it was in a submodule and thus had access to the non-public
internals of the struct.

We can remove the tests, because we don’t expose a way to specify
an error code to outside code anymore.

Suggested-by: Judson <nyarly@gmail.com>